New Compute API (#165)
This new API introduces new `Compute` API, which allows the value to *not* be updated
if needed, without being deleted. This saves the cost of updating the value
every time. This may seem small, but in a tight loop can have a pretty
significant impact.

More context can be found in https://github.com/puzpuzpuz/xsync/issues/160.