Cooperative rehashing in Map (#178)
Introduces cooperative rehashing for `xsync.Map` this means that goroutines that execute write operations, such as `Compute` or `Store`, may participate in table rehashing when the hash table grows or shrinks.

This behavior is always enabled, so the `WithSerialResize` function now acts as a no-op and is marked as deprecated.